Final Release v1.0

+++ app/src/main/java/com/example/memetemplate/ImageEditor.java
@@ -10,7 +10,6 @@ import android.annotation.SuppressLint;
 import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.content.pm.ActivityInfo;
 import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
@@ -75,8 +74,8 @@ public class ImageEditor extends AppCompatActivity implements BotttomSheet.Botto
                     tv_sticker.setControlItemsHidden(true);
                     Bitmap image = Bitmap.createBitmap(image_editor.getWidth(),  image_editor.getHeight(), Bitmap.Config.RGB_565);
                     image_editor.draw(new Canvas(image));
-                    int height = (int)(final_image.getHeight()*imView.getWidth()/final_image.getWidth());
-                    int top = (int)(imView.getHeight()/2 - height/2);
+                    int height = (final_image.getHeight()*imView.getWidth()/final_image.getWidth());
+                    int top = (imView.getHeight()/2 - height/2);
                     Bitmap temp = Bitmap.createBitmap(image, 0, top, image.getWidth(), height);
                     imView.setImageBitmap(temp);
                     final_image = temp;
